Create checkpoint 1b
Write a C++ program that takes a list of numbers and displays each one that is divisible by 3. Specifically, you should:
Write a function, getSize, that prompts for the size of the list, or the number of integers to be read in.
Write a function, getList, that accepts an array of integers and a count. It should prompt the user for the list of numbers and store them in the passed array.
Write a function, displayMultiples that accepts the array and its size. It should loop through each element in the array and display it if it is divisible by 3.
You may assume that the size of the list will always be less than 20.