Create mycheck2b.cpp
Write a C++ program according to the following:

Start with the skeleton of a file at: /home/cs165new/check02b.cpp
Create a struct for a Complex number that has a real part (a double) and an imaginary part (a double).
Write a function, prompt, to prompt the user for the real and imaginary parts of complex number, and have it return a Complex struct.
Write a function, display, that accepts a Complex struct and displays it in the format: "2.1 + 3.2i". You need not worry about handling zero or negatives at this point.
Fill in the body for the function, addComplex, that accepts two complex numbers and returns their sum. Remember that when adding complex numbers, the real parts are added and the imaginary parts are added.