Create MyAssign2.cpp
Write a C++ program according to the following:

Create a struct for an Access Record that contains a username, a filename, and a long (e.g. very large integer) timestamp.
Prompt the user for the name of a file.
Read in a list of Access Records from the file and store it in an array of structs.
Prompt the user for a start time and an end time for consideration.
Display a list of all files that were accessed during that period along with the users that accessed them.
For this assignment, you may assume that the information entered is all valid
For this assignment, you may assume that there will never be more than 500 lines in a file (so you can declare your array accordingly).
Don't forget to break your program into meaningful, cohesive functions!