Create myCheck03a.cpp
Write a C++ program that prompts the user for a positive even number, less than 100. If the user enters a number that is not valid, an exception message should be thrown from one function, caught in another and displayed to the user. In particular

Write a function, prompt, that prompts the user for a number, and returns it.
Have main call this function and then display the result.
In the prompt function, after getting the number from the user, the program should check for the following conditions (in this order) and throw a string with the accompanying message:
Condition: Number is less than 0. Message: "The number cannot be negative."
Condition: Number is greater than 100. Message: "The number cannot be greater than 100."
Condition: Number is odd. Message: "The number cannot be odd."
Set up main to try the function, and catch a potential exception.
If the exception is thrown. Main should display: "Error: " followed by the content of the exception message that was caught. It should not try display the number returned from the prompt function in this case.