Create myCheck04a.cpp
Create a class called "Book" with the following private data members:
title
author
Add a prompt method (function) to your class, which asks the user to enter a title and an author and stores them in the member variables. (Don't forget to use getline here to allow for spaces.)
Add a display method to your class, which displays the member variables in the format: "the book title" by the author.
For this assignment, you may use only one file and put the class and it's methods (functions) at the top of the file.
Have main:
Create a Book object.
Call its prompt method.
Call its display method.